First step towards created abstractions for future "tile" games.
The idea is each text game planned for this repository - tictactoe, chess, risk and battleship all revolve around two component parts - the board and the players.
The base unit of the board is each tile, of which a board is simply an array - a tile is a portion of the board that contains certain properties depending on the game. All games will employ
a common abstraction of this tile to hopefully simplify code between them.