Adding saving of games to persistent data. Considering its fairly useless to have a two player game that only functions on one client computer, the state of the game is saved in a file that can be accessed by multiple users across a machine. The idea is to host this game on a server computer, and have users (obviously known and trusted friends who have been given access before-hand) ssh into the server to play the game.
Each save file simply stores the .mark values of each tile in a string, a single digit indicating which player[index] is currently set to make their
turn, and info about the player that allows the program to know which player is which user. This would hopefully allow players to leave games and come back -
letting players play protracted games casually over the course of a day.