matchPlayers() function now returns correct player index matching with userid. However, I have now realised that the game is always saving a 0 over the opposing player's uid, as the game has no access to the opposing player's uid, and thus assumes it to be a 0.
While it is clunky and will be improved in the future, I will start writing a matchOtherPlayer function that will return the uid of the other player if found
in the save file. The returned uid will then be automatically saved as it should, leading to consistency in player ids between loads.