Optimized a small portion of the code - removing unecessary lines and variables. Fixed some formatting inconsistencies between files.
I realize now that the way updateBoard() works is very strange. It's purpose is to read the save file and pass the interpreted contents into the main boards().
At the time, I didn't realize that pointers transcended (to a certain extent) the restrictions on variable access between functions.
I should have made the function accept the pointer, make changes to it, and return a 0. Instead, the function creates its own pointer and returns the pointer -
I now know this is unecessary but I assumed at the time that pointers acted much like regular variables in functions.